Module rustfst::algorithms::arc_filters

Function objects to restrict which arcs are traversed in an FST.

Structs

AnyArcFilter
True for all arcs.

EpsilonArcFilter
True for (input/output) epsilon arcs.

InputEpsilonArcFilter
True for input epsilon arcs.

OutputEpsilonArcFilter
True for output epsilon arcs.

Traits

ArcFilter
Base trait to restrict which arcs are traversed in an FST.