Function rustfst::algorithms::concat

pub fn concat<W, F1, F2, F3>(fst_1: &F1, fst_2: &F2) -> Fallible<F3> where
    W: Semiring,
    F1: ExpandedFst<W = W>,
    F2: ExpandedFst<W = W>,
    F3: MutableFst<W = W>, 

Performs the concatenation of two wFSTs. If A transduces string x to y with weight a and B transduces string w to v with weight b, then their concatenation transduces string xw to yv with weight a ⊗ b.

Example

let fst_a : VectorFst<IntegerWeight> = fst![2 => 3];
let fst_b : VectorFst<IntegerWeight> = fst![6 => 5];

let fst_res : VectorFst<IntegerWeight> = concat(&fst_a, &fst_b)?;
let paths : HashSet<_> = fst_res.paths_iter().collect();

let mut paths_ref = HashSet::<FstPath<IntegerWeight>>::new();
paths_ref.insert(fst_path![2,6 => 3,5]);

assert_eq!(paths, paths_ref);