Function rustfst::algorithms::push_weights

pub fn push_weights<F>(
    fst: &mut F, 
    reweight_type: ReweightType, 
    remove_total_weight: bool
) -> Fallible<()> where
    F: Fst + ExpandedFst + MutableFst,
    F::W: WeaklyDivisibleSemiring,
    <<F as CoreFst>::W as Semiring>::ReverseWeight: 'static, 

Pushes the weights in FST in the direction defined by TYPE. If pushing towards the initial state, the sum of the weight of the outgoing transitions and final weight at a non-initial state is equal to One() in the resulting machine. If pushing towards the final state, the same property holds on the reverse machine.