Struct rustfst::fst_impls::VectorFst

pub struct VectorFst<W: Semiring> { /* fields omitted */ }

Simple concrete, mutable FST whose states and arcs are stored in standard vectors.

All states are stored in a vector of states. In each state, there is a vector of arcs containing the outgoing transitions.