Module rustfst::fst_properties

Provides the FstProperties struct and some utils functions around it. Useful to assert some properties on a Fst.

Modules

mutable_properties
Functions for getting property bit vectors when executing mutation operations.

Structs

FstProperties
The property bits here assert facts about an FST. If individual bits are added, then the composite fst_properties below, the property functions and property names in fst_properties.cc, and TestProperties() in test-fst_properties.h should be updated. For each of these fst_properties below there is a pair of property bits, one positive and one negative. If the positive bit is set, the property is true. If the negative bit is set, the property is false. If neither is set, the property has unknown value. Both should never be simultaneously set. The individual positive and negative bit pairs should be adjacent with the positive bit at an odd and lower position.

Functions

compat_properties
Tests compatibility between two sets of properties.

compute_fst_properties
Computes all the FstProperties of the FST bit don't attach them to the FST.

known_properties
Both bits are set iff one bit is set.