Struct rustfst::fst_properties::FstProperties

pub struct FstProperties { /* fields omitted */ }

The property bits here assert facts about an FST. If individual bits are added, then the composite fst_properties below, the property functions and property names in fst_properties.cc, and TestProperties() in test-fst_properties.h should be updated. For each of these fst_properties below there is a pair of property bits, one positive and one negative. If the positive bit is set, the property is true. If the negative bit is set, the property is false. If neither is set, the property has unknown value. Both should never be simultaneously set. The individual positive and negative bit pairs should be adjacent with the positive bit at an odd and lower position.

Methods

impl FstProperties

pub const ACCEPTOR: FstProperties

ilabel == olabel for each arc.

pub const NOT_ACCEPTOR: FstProperties

ilabel != olabel for some arc.

pub const I_DETERMINISTIC: FstProperties

ilabels unique leaving each state.

pub const NOT_I_DETERMINISTIC: FstProperties

ilabels not unique leaving some state.

pub const O_DETERMINISTIC: FstProperties

olabels unique leaving each state.

pub const NOT_O_DETERMINISTIC: FstProperties

olabels not unique leaving some state.

pub const EPSILONS: FstProperties

FST has input/output epsilons.

pub const NO_EPSILONS: FstProperties

FST has no input/output epsilons.

pub const I_EPSILONS: FstProperties

FST has input epsilons.

pub const NO_I_EPSILONS: FstProperties

FST has no input epsilons.

pub const O_EPSILONS: FstProperties

FST has output epsilons.

pub const NO_O_EPSILONS: FstProperties

FST has no output epsilons.

pub const I_LABEL_SORTED: FstProperties

ilabels sorted wrt < for each state.

pub const NOT_I_LABEL_SORTED: FstProperties

ilabels not sorted wrt < for some state.

pub const O_LABEL_SORTED: FstProperties

olabels sorted wrt < for each state.

pub const NOT_O_LABEL_SORTED: FstProperties

olabels not sorted wrt < for some state.

pub const WEIGHTED: FstProperties

Non-trivial arc or final weights.

pub const UNWEIGHTED: FstProperties

Only trivial arc and final weights.

pub const CYCLIC: FstProperties

FST has cycles.

pub const ACYCLIC: FstProperties

FST has no cycles.

pub const INITIAL_CYCLIC: FstProperties

FST has cycles containing the initial state.

pub const INITIAL_ACYCLIC: FstProperties

FST has no cycles containing the initial state.

pub const TOP_SORTED: FstProperties

FST is topologically sorted.

pub const NOT_TOP_SORTED: FstProperties

FST is not topologically sorted.

pub const ACCESSIBLE: FstProperties

All states reachable from the initial state.

pub const NOT_ACCESSIBLE: FstProperties

Not all states reachable from the initial state.

pub const COACCESSIBLE: FstProperties

All states can reach a final state.

pub const NOT_COACCESSIBLE: FstProperties

Not all states can reach a final state.

pub const STRING: FstProperties

If NumStates() > 0, then state 0 is initial, state NumStates() - 1 is final, there is a transition from each non-final state i to state i + 1, and there are no other transitions.

pub const NOT_STRING: FstProperties

Not a string FST.

pub const WEIGHTED_CYCLES: FstProperties

FST has at least one weighted cycle.

pub const UNWEIGHTED_CYCLES: FstProperties

Only unweighted cycles.

pub const NULL_PROPERTIES: FstProperties

Properties of an empty machine.

pub const COPY_PROPERTIES: FstProperties

Properties that are preserved when an FST is copied.

pub const INTRINSIC_PROPERTIES: FstProperties

Properties that are intrinsic to the FST.

pub const SET_START_PROPERTIES: FstProperties

Properties that are preserved when an FST start state is set.

pub const SET_FINAL_PROPERTIES: FstProperties

Properties that are preserved when an FST final weight is set.

pub const ADD_STATE_PROPERTIES: FstProperties

Properties that are preserved when an FST state is added.

pub const ADD_ARC_PROPERTIES: FstProperties

Properties that are preserved when an FST arc is added.

pub const SET_ARC_PROPERTIES: FstProperties

Properties that are preserved when an FST arc is set.

pub const DELETE_STATES_PROPERTIES: FstProperties

Properties that are preserved when FST states are deleted.

pub const DELETE_ARCS_PROPERTIES: FstProperties

Properties that are preserved when FST arcs are deleted.

pub const STATESORT_PROPERTIES: FstProperties

Properties that are preserved when an FST's states are reordered.

pub const ARCSORT_PROPERTIES: FstProperties

Properties that are preserved when an FST's arcs are reordered.

pub const I_LABEL_INVARIANT_PROPERTIES: FstProperties

Properties that are preserved when an FST's input labels are changed.

pub const O_LABEL_INVARIANT_PROPERTIES: FstProperties

Properties that are preserved when an FST's output labels are changed.

pub const WEIGHT_INVARIANT_PROPERTIES: FstProperties

Properties that are preserved when an FST's weights are changed. This assumes that the set of states that are non-final is not changed.

pub const ADD_SUPER_FINAL_PROPERTIES: FstProperties

Properties that are preserved when a superfinal state is added and an FST's final weights are directed to it via new transitions.

pub const RM_SUPER_FINAL_PROPERTIES: FstProperties

Properties that are preserved when a superfinal state is removed and the epsilon transitions directed to it are made final weights.

pub const POS_PROPERTIES: FstProperties

pub const NEG_PROPERTIES: FstProperties

pub const ALL_PROPERTIES: FstProperties

pub fn empty() -> FstProperties

Returns an empty set of flags.

pub fn all() -> FstProperties

Returns the set containing all flags.

pub fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub fn from_bits(bits: u32) -> Option<FstProperties>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub fn from_bits_truncate(bits: u32) -> FstProperties

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub fn intersects(&self, other: FstProperties) -> bool

Returns true if there are flags common to both self and other.

pub fn contains(&self, other: FstProperties) -> bool

Returns true all of the flags in other are contained within self.

pub fn insert(&mut self, other: FstProperties)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: FstProperties)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: FstProperties)

Toggles the specified flags in-place.

pub fn set(&mut self, other: FstProperties, value: bool)

Inserts or removes the specified flags depending on the passed value.