Module rustfst::fst_traits

Provides traits that must be implemented to be able to use generic algorithms.

Traits

ArcIterator
Trait to iterate over the outgoing arcs of a particular state in a wFST

BinaryDeserializer

BinarySerializer

CoreFst
Trait defining necessary methods for a wFST to access start states and final states.

ExpandedFst
Trait defining the necessary methods that should implement an ExpandedFST e.g a FST where all the states are already computed and not computed on the fly.

FinalStatesIterator
Trait to iterate over the final states of a wFST.

Fst
Trait defining the minimum interface necessary for a wFST.

MutableArcIterator
Iterate over mutable arcs in a wFST.

MutableFst
Trait defining the methods to modify a wFST.

PathsIterator
Trait to iterate over the paths accepted by an FST.

StateIterator
Trait to iterate over the states of a wFST.

TextParser
Trait to allow serialization and deserialization of a wFST in text format.