Trait rustfst::fst_traits::ArcIterator

pub trait ArcIterator<'a>: CoreFst where
    Self::W: 'a, {
    type Iter: Iterator<Item = &'a Arc<Self::W>> + Clone;
    fn arcs_iter(&'a self, state_id: StateId) -> Fallible<Self::Iter>;

    unsafe fn arcs_iter_unchecked(&'a self, state_id: StateId) -> Self::Iter;
}

Trait to iterate over the outgoing arcs of a particular state in a wFST

Associated Types

type Iter: Iterator<Item = &'a Arc<Self::W>> + Clone

Iterator used to iterate over the arcs leaving a state of an FST.

Loading content...
Required methods

fn arcs_iter(&'a self, state_id: StateId) -> Fallible<Self::Iter>

unsafe fn arcs_iter_unchecked(&'a self, state_id: StateId) -> Self::Iter

Loading content...