Trait rustfst::fst_traits::CoreFst

pub trait CoreFst {
    type W: Semiring;
    fn start(&self) -> Option<StateId>;

    fn final_weight(&self, state_id: StateId) -> Fallible<Option<&Self::W>>;

    unsafe fn final_weight_unchecked(
        &self, 
        state_id: StateId
    ) -> Option<&Self::W>;

    fn num_arcs(&self, s: StateId) -> Fallible<usize>;

    unsafe fn num_arcs_unchecked(&self, s: StateId) -> usize;

    fn is_final(&self, state_id: StateId) -> Fallible<bool> { ... }

    unsafe fn is_final_unchecked(&self, state_id: StateId) -> bool { ... }

    fn is_start(&self, state_id: StateId) -> bool { ... }
}

Trait defining necessary methods for a wFST to access start states and final states.

Associated Types

type W: Semiring

Weight use in the wFST. This type must implement the Semiring trait.

Loading content...
Required methods

fn start(&self) -> Option<StateId>

Returns the ID of the start state of the wFST if it exists else none.

Example

// 1 - Create an FST
let mut fst = VectorFst::<BooleanWeight>::new();
let s = fst.add_state();
fst.set_start(s);

// 2 - Access the start state
let start_state = fst.start();
assert_eq!(start_state, Some(s));

fn final_weight(&self, state_id: StateId) -> Fallible<Option<&Self::W>>

Retrieves the final weight of a state (if the state is a final one).

Example

// 1 - Create an FST
let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();
fst.set_final(s2, BooleanWeight::one());

// 2 - Access the final weight of each state
assert_eq!(fst.final_weight(s1).unwrap(), None);
assert_eq!(fst.final_weight(s2).unwrap(), Some(&BooleanWeight::one()));
assert!(fst.final_weight(s2 + 1).is_err());

unsafe fn final_weight_unchecked(&self, state_id: StateId) -> Option<&Self::W>

fn num_arcs(&self, s: StateId) -> Fallible<usize>

Number of arcs leaving a specific state in the wFST.

Example

let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();

assert_eq!(fst.num_arcs(s1).unwrap(), 0);
fst.add_arc(s1, Arc::new(3, 5, BooleanWeight::new(true), s2));
assert_eq!(fst.num_arcs(s1).unwrap(), 1);

unsafe fn num_arcs_unchecked(&self, s: StateId) -> usize

Loading content...
Provided methods

fn is_final(&self, state_id: StateId) -> Fallible<bool>

Returns whether or not the state with identifier passed as parameters is a final state.

Example

// 1 - Create an FST
let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();
fst.set_final(s2, BooleanWeight::one());

// 2 - Test if a state is final
assert_eq!(fst.is_final(s1).unwrap(), false);
assert_eq!(fst.is_final(s2).unwrap(), true);
assert!(fst.is_final(s2 + 1).is_err());

unsafe fn is_final_unchecked(&self, state_id: StateId) -> bool

fn is_start(&self, state_id: StateId) -> bool

Check whether a state is the start state or not.

Loading content...