Trait rustfst::fst_traits::ExpandedFst

pub trait ExpandedFst: Fst {
    fn num_states(&self) -> usize;

    fn write_text<P: AsRef<Path>>(&self, path_output: P) -> Fallible<()> { ... }

    fn text(&self) -> Fallible<String> { ... }

    fn draw<P: AsRef<Path>>(
        &self, 
        path_output: P, 
        config: &DrawingConfig
    ) -> Fallible<()> { ... }

    fn properties(&self) -> Fallible<FstProperties> { ... }
}

Trait defining the necessary methods that should implement an ExpandedFST e.g a FST where all the states are already computed and not computed on the fly.

Required methods

fn num_states(&self) -> usize

Returns the number of states that contains the FST. They are all counted even if some states are not on a successful path (doesn't perform triming).

Example

let mut fst = VectorFst::<BooleanWeight>::new();

assert_eq!(fst.num_states(), 0);

fst.add_state();
assert_eq!(fst.num_states(), 1);

fst.add_state();
assert_eq!(fst.num_states(), 2);

Loading content...
Provided methods

fn write_text<P: AsRef<Path>>(&self, path_output: P) -> Fallible<()>

Serializes the FST as a text file in a format compatible with OpenFST.

fn text(&self) -> Fallible<String>

Writes the text representation of the FST into a String.

fn draw<P: AsRef<Path>>(
&self,
path_output: P,
config: &DrawingConfig
) -> Fallible<()>

Serializes the FST as a DOT file compatible with GraphViz binaries.

fn properties(&self) -> Fallible<FstProperties>

Compute the properties verified by the Fst.

Loading content...