Trait rustfst::fst_traits::Fst

pub trait Fst: CoreFst + PartialEq + Clone + for<'a> ArcIterator<'a> + for<'b> StateIterator<'b> + Display + Debug {
    fn num_input_epsilons(&self, state: StateId) -> Fallible<usize> { ... }

    fn num_output_epsilons(&self, state: StateId) -> Fallible<usize> { ... }

    fn is_acceptor(&self) -> bool { ... }
}

Trait defining the minimum interface necessary for a wFST.

Provided methods

fn num_input_epsilons(&self, state: StateId) -> Fallible<usize>

Returns the number of arcs with epsilon input labels leaving a state.

Example :

let mut fst = VectorFst::new();
let s0 = fst.add_state();
let s1 = fst.add_state();

fst.add_arc(s0, Arc::new(EPS_LABEL, 18, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(76, EPS_LABEL, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(EPS_LABEL, 18, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(45, 18, IntegerWeight::one(), s0));
fst.add_arc(s1, Arc::new(76, 18, IntegerWeight::one(), s1));

assert_eq!(fst.num_input_epsilons(s0).unwrap(), 2);
assert_eq!(fst.num_input_epsilons(s1).unwrap(), 0);

fn num_output_epsilons(&self, state: StateId) -> Fallible<usize>

Returns the number of arcs with epsilon output labels leaving a state.

Example :

let mut fst = VectorFst::new();
let s0 = fst.add_state();
let s1 = fst.add_state();

fst.add_arc(s0, Arc::new(EPS_LABEL, 18, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(76, EPS_LABEL, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(EPS_LABEL, 18, IntegerWeight::one(), s1));
fst.add_arc(s0, Arc::new(45, 18, IntegerWeight::one(), s0));
fst.add_arc(s1, Arc::new(76, 18, IntegerWeight::one(), s1));

assert_eq!(fst.num_output_epsilons(s0).unwrap(), 1);
assert_eq!(fst.num_output_epsilons(s1).unwrap(), 0);

fn is_acceptor(&self) -> bool

Returns true if the Fst is an acceptor. False otherwise. Acceptor means for all arc, arc.ilabel == arc.olabel

Loading content...