Trait rustfst::fst_traits::MutableFst

pub trait MutableFst: Fst + for<'a> MutableArcIterator<'a> {
    fn new() -> Self;

    fn set_start(&mut self, state_id: StateId) -> Fallible<()>;

    unsafe fn set_start_unchecked(&mut self, state_id: StateId);

    fn set_final(
        &mut self, 
        state_id: StateId, 
        final_weight: Self::W
    ) -> Fallible<()>;

    unsafe fn set_final_unchecked(
        &mut self, 
        state_id: StateId, 
        final_weight: Self::W
    );

    fn add_state(&mut self) -> StateId;

    fn add_states(&mut self, n: usize);

    fn del_state(&mut self, state_id: StateId) -> Fallible<()>;

    fn del_states<T: IntoIterator<Item = StateId>>(
        &mut self, 
        states: T
    ) -> Fallible<()>;

    unsafe fn del_arcs_id_sorted_unchecked(
        &mut self, 
        state: StateId, 
        to_del: &Vec<usize>
    );

    fn add_arc(&mut self, source: StateId, arc: Arc<Self::W>) -> Fallible<()>;

    unsafe fn add_arc_unchecked(&mut self, source: StateId, arc: Arc<Self::W>);

    unsafe fn set_arcs_unchecked(
        &mut self, 
        source: StateId, 
        arcs: Vec<Arc<Self::W>>
    );

    fn delete_final_weight(&mut self, source: StateId) -> Fallible<()>;

    fn delete_arcs(&mut self, source: StateId) -> Fallible<()>;

    fn pop_arcs(&mut self, source: StateId) -> Fallible<Vec<Arc<Self::W>>>;

    unsafe fn pop_arcs_unchecked(
        &mut self, 
        source: StateId
    ) -> Vec<Arc<Self::W>>;

    fn reserve_arcs(
        &mut self, 
        source: StateId, 
        additional: usize
    ) -> Fallible<()>;

    unsafe fn reserve_arcs_unchecked(
        &mut self, 
        source: StateId, 
        additional: usize
    );

    fn reserve_states(&mut self, additional: usize);

    fn final_weight_mut(
        &mut self, 
        state_id: StateId
    ) -> Fallible<Option<&mut Self::W>>;

    unsafe fn final_weight_unchecked_mut(
        &mut self, 
        state_id: StateId
    ) -> Option<&mut Self::W>;

    fn sort_arcs_unchecked<F: Fn(&Arc<Self::W>, &Arc<Self::W>) -> Ordering>(
        &mut self, 
        state: StateId, 
        f: F
    );

    unsafe fn unique_arcs_unchecked(&mut self, state: StateId);

    unsafe fn sum_arcs_unchecked(&mut self, state: StateId);

    fn add_fst<F: ExpandedFst<W = Self::W>>(
        &mut self, 
        fst_to_add: &F
    ) -> Fallible<HashMap<StateId, StateId>> { ... }

    fn closure_plus(&mut self) { ... }

    fn closure_star(&mut self) { ... }

    fn arc_map<M: ArcMapper<Self::W>>(&mut self, mapper: &mut M) -> Fallible<()> { ... }
}

Trait defining the methods to modify a wFST.

Required methods

fn new() -> Self

Creates an empty wFST.

fn set_start(&mut self, state_id: StateId) -> Fallible<()>

The state with identifier state_id is now the start state. Note that only one start state is allowed in this implementation. Calling this function twice will mean losing the first start state. If the state_id doesn't exist an error is raised.

let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();

assert_eq!(fst.start(), None);

fst.set_start(s1);
assert_eq!(fst.start(), Some(s1));

fst.set_start(s2);
assert_eq!(fst.start(), Some(s2));

unsafe fn set_start_unchecked(&mut self, state_id: StateId)

fn set_final(
&mut self,
state_id: StateId,
final_weight: Self::W
) -> Fallible<()>

The state with identifier state_id is now a final state with a weight final_weight. If the state_id doesn't exist an error is raised.

let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();

assert_eq!(fst.final_weight(s1).unwrap(), None);
assert_eq!(fst.final_weight(s2).unwrap(), None);

fst.set_final(s1, BooleanWeight::one());
assert_eq!(fst.final_weight(s1).unwrap(), Some(&BooleanWeight::one()));
assert_eq!(fst.final_weight(s2).unwrap(), None);

fst.set_final(s2, BooleanWeight::one());
assert_eq!(fst.final_weight(s1).unwrap(), Some(&BooleanWeight::one()));
assert_eq!(fst.final_weight(s2).unwrap(), Some(&BooleanWeight::one()));

unsafe fn set_final_unchecked(
&mut self,
state_id: StateId,
final_weight: Self::W
)

fn add_state(&mut self) -> StateId

Adds a new state to the current FST. The identifier of the new state is returned

Example

let mut fst = VectorFst::<BooleanWeight>::new();

assert_eq!(fst.num_states(), 0);

fst.add_state();
assert_eq!(fst.num_states(), 1);

fst.add_state();
assert_eq!(fst.num_states(), 2);

fn add_states(&mut self, n: usize)

fn del_state(&mut self, state_id: StateId) -> Fallible<()>

Removes a state from an FST. It also removes all the arcs starting from another state and reaching this state. An error is raised if the state state_id doesn't exist.

Example

let mut fst = VectorFst::<BooleanWeight>::new();

assert_eq!(fst.states_iter().count(), 0);

let s1 = fst.add_state();

assert_eq!(fst.states_iter().count(), 1);

fst.del_state(s1);

assert_eq!(fst.states_iter().count(), 0);

fn del_states<T: IntoIterator<Item = StateId>>(
&mut self,
states: T
) -> Fallible<()>

Removes multiple states from an FST. If one of the states doesn't exist, an error is raised.

Warning

This method modifies the id of the states that are left in the FST. Id that were used before calling this function should no longer be used.

Example

let mut fst = VectorFst::<BooleanWeight>::new();

assert_eq!(fst.states_iter().count(), 0);

let s1 = fst.add_state();
let s2 = fst.add_state();

assert_eq!(fst.states_iter().count(), 2);

let states_to_remove = vec![s1, s2];
fst.del_states(states_to_remove.into_iter());

assert_eq!(fst.states_iter().count(), 0);

unsafe fn del_arcs_id_sorted_unchecked(
&mut self,
state: StateId,
to_del: &Vec<usize>
)

fn add_arc(&mut self, source: StateId, arc: Arc<Self::W>) -> Fallible<()>

Adds an arc to the FST. The arc will start in the state source. An error is raised if the state source doesn't exist.

Warning

This method modifies the id of the states that are left in the FST. Id that were used before calling this function should no longer be used.

Example

let mut fst = VectorFst::<BooleanWeight>::new();
let s1 = fst.add_state();
let s2 = fst.add_state();

assert_eq!(fst.num_arcs(s1).unwrap(), 0);
fst.add_arc(s1, Arc::new(3, 5, BooleanWeight::new(true), s2));
assert_eq!(fst.num_arcs(s1).unwrap(), 1);

unsafe fn add_arc_unchecked(&mut self, source: StateId, arc: Arc<Self::W>)

unsafe fn set_arcs_unchecked(
&mut self,
source: StateId,
arcs: Vec<Arc<Self::W>>
)

fn delete_final_weight(&mut self, source: StateId) -> Fallible<()>

Remove the final weight of a specific state.

fn delete_arcs(&mut self, source: StateId) -> Fallible<()>

Deletes all the arcs leaving a state.

fn pop_arcs(&mut self, source: StateId) -> Fallible<Vec<Arc<Self::W>>>

Remove all arcs leaving a state and return them.

unsafe fn pop_arcs_unchecked(&mut self, source: StateId) -> Vec<Arc<Self::W>>

fn reserve_arcs(&mut self, source: StateId, additional: usize) -> Fallible<()>

Reserve space for storing enough arcs leaving a state.

unsafe fn reserve_arcs_unchecked(&mut self, source: StateId, additional: usize)

fn reserve_states(&mut self, additional: usize)

Reserve space for storing enough states.

fn final_weight_mut(
&mut self,
state_id: StateId
) -> Fallible<Option<&mut Self::W>>

Retrieves a mutable reference to the final weight of a state (if the state is a final one).

unsafe fn final_weight_unchecked_mut(
&mut self,
state_id: StateId
) -> Option<&mut Self::W>

fn sort_arcs_unchecked<F: Fn(&Arc<Self::W>, &Arc<Self::W>) -> Ordering>(
&mut self,
state: StateId,
f: F
)

unsafe fn unique_arcs_unchecked(&mut self, state: StateId)

unsafe fn sum_arcs_unchecked(&mut self, state: StateId)

Loading content...
Provided methods

fn add_fst<F: ExpandedFst<W = Self::W>>(
&mut self,
fst_to_add: &F
) -> Fallible<HashMap<StateId, StateId>>

fn closure_plus(&mut self)

This operation computes the concatenative closure. If A transduces string x to y with weight a, then the closure transduces x to y with weight a, xx to yy with weight a ⊗ a, xxx to yyy with weight a ⊗ a ⊗ a, etc.

fn closure_star(&mut self)

This operation computes the concatenative closure. If A transduces string x to y with weight a, then the closure transduces x to y with weight a, xx to yy with weight a ⊗ a, xxx to yyy with weight a ⊗ a ⊗ a, etc. The empty string is transduced to itself with weight 1 as well.

fn arc_map<M: ArcMapper<Self::W>>(&mut self, mapper: &mut M) -> Fallible<()>

Maps an arc using a ArcMapper object.

Loading content...