Trait rustfst::semirings::Semiring

pub trait Semiring: Clone + PartialEq + PartialOrd + Debug + Default + Display + AsRef<Self> + Hash + Eq + Sized {
    type Type: Clone;
    type ReverseWeight: Semiring;
    fn zero() -> Self;

    fn one() -> Self;

    fn new(value: Self::Type) -> Self;

    fn plus_assign<P: AsRef<Self>>(&mut self, rhs: P) -> Fallible<()>;

    fn times_assign<P: AsRef<Self>>(&mut self, rhs: P) -> Fallible<()>;

    fn value(&self) -> &Self::Type;

    fn take_value(self) -> Self::Type;

    fn set_value(&mut self, value: Self::Type);

    fn reverse(&self) -> Fallible<Self::ReverseWeight>;

    fn properties() -> SemiringProperties;

    fn plus<P: AsRef<Self>>(&self, rhs: P) -> Fallible<Self> { ... }

    fn times<P: AsRef<Self>>(&self, rhs: P) -> Fallible<Self> { ... }

    fn is_one(&self) -> bool { ... }

    fn is_zero(&self) -> bool { ... }
}

For some operations, the weight set associated to a wFST must have the structure of a semiring. (S, +, *, 0, 1) is a semiring if (S, +, 0) is a commutative monoid with identity element 0, (S, *, 1) is a monoid with identity element 1, * distributes over +, 0 is an annihilator for *. Thus, a semiring is a ring that may lack negation. For more information : https://cs.nyu.edu/~mohri/pub/hwa.pdf

Associated Types

type Type: Clone

type ReverseWeight: Semiring

Loading content...
Required methods

fn zero() -> Self

fn one() -> Self

fn new(value: Self::Type) -> Self

fn plus_assign<P: AsRef<Self>>(&mut self, rhs: P) -> Fallible<()>

fn times_assign<P: AsRef<Self>>(&mut self, rhs: P) -> Fallible<()>

fn value(&self) -> &Self::Type

Borrow underneath value.

fn take_value(self) -> Self::Type

Move underneath value.

fn set_value(&mut self, value: Self::Type)

fn reverse(&self) -> Fallible<Self::ReverseWeight>

fn properties() -> SemiringProperties

Loading content...
Provided methods

fn plus<P: AsRef<Self>>(&self, rhs: P) -> Fallible<Self>

fn times<P: AsRef<Self>>(&self, rhs: P) -> Fallible<Self>

fn is_one(&self) -> bool

fn is_zero(&self) -> bool

Loading content...